+++ common/src/main/java/immersive_furniture/item/FurnitureItem.java
@@ -2,19 +2,24 @@ package immersive_furniture.item;
 
 import immersive_furniture.Blocks;
 import immersive_furniture.data.FurnitureData;
+import immersive_furniture.data.ServerFurnitureRegistry;
 import net.minecraft.ChatFormatting;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.network.chat.Component;
+import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.item.BlockItem;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.TooltipFlag;
+import net.minecraft.world.item.context.BlockPlaceContext;
 import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.state.BlockState;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.List;
 
 public class FurnitureItem extends BlockItem {
     public static final String FURNITURE = "Furniture";
+    public static final String FURNITURE_HASH = "FurnitureHash";
 
     public FurnitureItem(Properties settings) {
         super(Blocks.FURNITURE.get(), settings);
@@ -42,5 +47,15 @@ public class FurnitureItem extends BlockItem {
     public static void setData(ItemStack stack, FurnitureData data) {
         CompoundTag tag = stack.getOrCreateTagElement(BLOCK_ENTITY_TAG);
         tag.put(FURNITURE, data.toTag());
+    }
+
+    @Override
+    protected boolean placeBlock(BlockPlaceContext context, BlockState state) {
+        // Keep track of placed furniture to estimate usage
+        if (context.getLevel() instanceof ServerLevel level) {
+            ServerFurnitureRegistry.increase(level, getData(context.getItemInHand()));
+        }
+
+        return super.placeBlock(context, state);
     }
 }