+++ fabric/src/main/java/immersive_furniture/fabric/CommonFabric.java
@@ -1,10 +1,12 @@
 package immersive_furniture.fabric;
 
 import immersive_furniture.*;
+import immersive_furniture.data.ServerFurnitureRegistry;
 import immersive_furniture.fabric.cobalt.network.NetworkHandlerImpl;
 import immersive_furniture.fabric.cobalt.registration.RegistrationImpl;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.itemgroup.v1.FabricItemGroup;
+import net.fabricmc.fabric.api.networking.v1.ServerPlayConnectionEvents;
 import net.minecraft.core.Registry;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.world.item.CreativeModeTab;
@@ -23,6 +25,9 @@ public final class CommonFabric implements ModInitializer {
         Sounds.bootstrap();
 
         Messages.loadMessages();
+
+        ServerPlayConnectionEvents.JOIN.register((player, handler, sender) ->
+                ServerFurnitureRegistry.syncWithPlayer(player.player));
 
         CreativeModeTab group = FabricItemGroup.builder()
                 .title(ItemGroups.getDisplayName())